backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
google-genai>=1.0.0
python-multipart>=0.0.9
Pillow>=10.0.0
python-dotenv>=1.0.1
